Added YML to CSV and CSV to YML converters, documentation, and an example
This is useful because we can edit data in excel, which is a great
interface, and then export to CSV, and use the script to convert to YML.
Additionally, for existing YML data, we can use the reverse converter to
get everything into CSV, and hence, Excel.

See README for documentation.